BookChain/requirements.txt
Flask<1
flask-bootstrap4
flask-wtf
flask-testing
blinker
firebase-admin
flask-login